requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
textblob==0.17.1
gunicorn==21.2.0